model.*.runtime must not use lazy activation
The model.*.runtime bundles must be started automatically, we need them
running for ESH.
The lazy bundle activation will retain the execution of the activator's
start method until a class is accessed first. This will result in not
started bundles and missing functionality (e.g. a bundle does not export
a package or no one use an exported package).

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=477075
Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>